pyproject.toml
[project]
name = "siteplan-lite"
version = "0.1.0"
description = "Construction Management Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "starlette>=0.46.2",
    "stringgenerator>=0.4.4",
    "jinja2>=3.1.6",
    "pydantic[email]>=2.11.3",
    "starlette-wtf>=0.4.5",
    "uvicorn>=0.34.2",
    "starlette-htmx>=0.1.1",
    "starlette-login>=1.0.2",
    "werkzeug>=3.1.3",
    "httpx>=0.28.1",
    "pydantic-extra-types>=2.10.3",
    "pycountry>=24.6.1",
    "phonenumbers>=9.0.3",
    "starlette-decorouter>=1.0.0",
    "starlette-compress>=1.5.1",
    "shiny>=1.4.0",
    "shinywidgets>=0.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
